+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/jws/SdJwtSerializationTest.kt
@@ -7,9 +7,12 @@ import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldContain
 import io.kotest.matchers.string.shouldNotContain
+import io.kotest.matchers.types.shouldBeInstanceOf
 import io.matthewnelson.encoding.base64.Base64
 import io.matthewnelson.encoding.core.Decoder.Companion.decodeToByteArray
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
+import kotlinx.serialization.json.JsonArray
+import kotlinx.serialization.json.jsonObject
 import kotlin.random.Random
 import kotlin.random.nextUInt
 
@@ -114,6 +117,39 @@ class SdJwtSerializationTest : FreeSpec({
         disclosure shouldBe "WyJfMjZiYzRMVC1hYzZxMktJNmNCVzVlcyIsImZhbWlseV9uYW1lIiwiTcO2Yml1cyJd"
 
         "WyJsa2x4RjVqTVlsR1RQVW92TU5JdkNBIiwgIkZSIl0".hashDisclosure() shouldBe "w0I8EKcdCtUPkGCNUrfwVp2xEgNjtoIDlOxc9-PlOhs"
+    }
+
+    "Serialize nested Byte Array" {
+        val salt = Random.nextBytes(32)
+        val name = Random.nextBytes(16).encodeToString(Base64())
+        val value = listOf(Random.nextBytes(16))
+        val item = SelectiveDisclosureItem(salt, name, value)
+
+        val serialized = item.serialize()
+
+        serialized shouldContain "["
+        serialized shouldContain """"${salt.encodeToString(Base64UrlStrict)}""""
+        serialized shouldContain """"$name""""
+        serialized shouldContain """"${value.first().encodeToString(Base64UrlStrict)}""""
+        serialized shouldContain "]"
+
+        val deserialized = SelectiveDisclosureItem.deserialize(serialized).getOrThrow()
+        deserialized shouldBe item
+    }
+
+    "Deserialize nested from spec" {
+        val input = "WyJsa2x4RjVqTVlsR1RQVW92TU5JdkNBIiwgImFkZHJlc3MiLCB7Il9zZCI6IFs" +
+                "iWEZjN3pYUG03enpWZE15d20yRXVCZmxrYTVISHF2ZjhVcF9zek5HcXZpZyIsICJiZDF" +
+                "FVnpnTm9wVWs0RVczX2VRMm4zX05VNGl1WE9IdjlYYkdITjNnMVRFIiwgImZfRlFZZ3Z" +
+                "RV3Z5VnFObklYc0FSbE55ZTdZR3A4RTc3Z1JBamFxLXd2bnciLCAidjRra2JfcFAxamx" +
+                "2VWJTanR5YzVicWNXeUEtaThYTHZoVllZN1pUMHRiMCJdfV0"
+
+        val decoded = input.decodeToByteArray(Base64()).decodeToString()
+        val deserialized = SelectiveDisclosureItem.deserialize(decoded).getOrThrow()
+        deserialized.claimName shouldBe "address"
+        val nestedSd = deserialized.claimValue.jsonObject["_sd"]
+            .shouldBeInstanceOf<JsonArray>()
+        nestedSd.size shouldBe 4
     }
 
 })